requirements.txt
# AWS Infrastruct - Complete Requirements File
# Generated by Kiro on 2025-01-11

# Core AI/ML Framework
strands-agents>=1.0.0
strands-tools>=1.0.0

# AWS SDK and Cloud Services
boto3>=1.34.0
botocore>=1.34.0

# Environment and Configuration Management
python-dotenv>=1.0.0

# Data Processing and Serialization
PyYAML>=6.0.1
pyyaml>=6.0.1

# GUI Framework (Built into Python, but listed for clarity)
# tkinter - Built into Python standard library

# Database Support
# sqlite3 - Built into Python standard library

# Enhanced Diagram Generation (Optional - for enhanced GUI)
diagrams>=0.23.0
graphviz>=0.20.0
Pillow>=10.0.0

# Model Context Protocol (MCP) Support
mcp>=1.0.0

# Browser Automation (Optional - for enhanced deployment)
# nova-act - Install separately with: python install_nova_act.py
# Note: nova-act may require Visual Studio Build Tools on Windows

# Web and HTTP Support (Built into Python)
# webbrowser - Built into Python standard library
# tempfile - Built into Python standard library
# json - Built into Python standard library
# base64 - Built into Python standard library

# System and OS Support (Built into Python)
# os - Built into Python standard library
# sys - Built into Python standard library
# subprocess - Built into Python standard library
# threading - Built into Python standard library
# datetime - Built into Python standard library
# re - Built into Python standard library

# Text Processing and Utilities (Built into Python)
# typing - Built into Python standard library (Python 3.5+)

# Development and Testing (Optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Additional Dependencies for Enhanced Features
# These are optional and the application will work without them
# with graceful fallbacks

# For enhanced templating
# jinja2>=3.1.0  # Optional - for advanced template processing

# For enhanced logging
# loguru>=0.7.0  # Optional - for better logging

# For enhanced error handling
# sentry-sdk>=1.0.0  # Optional - for error tracking

# Platform-specific notes:
# Windows: May need Visual Studio Build Tools for some packages
# macOS: May need Xcode Command Line Tools
# Linux: May need build-essential package

# Installation order recommendation:
# 1. pip install --upgrade pip
# 2. pip install -r requirements.txt
# 3. python install_nova_act.py (optional, for enhanced deployment)
# 4. Configure .env file with AWS credentials